Module WParserAndCompiler.Annot

type nest_kind =

| LoopBody of string
(*
This command nests its loop body an (abstracted) function call
*)

| FunCall of string
(*
This command nests the body of a function call
*)

val nest_kind_to_yojson : nest_kind -> Yojson.Safe.t

val nest_kind_of_yojson : Yojson.Safe.t -> nest_kind Ppx_deriving_yojson_runtime.error_or

type stmt_kind =

| Normal of bool
(*
A command that makes up part of a WISL statement, and whether this is the last cmd of said statement
*)

| Return of bool
(*
Same as Normal, but specific to the return statement
*)

| Hidden
(*
A command that doesn't map to a particular WISL statement
*)

| LoopPrefix
(*
A command in the prefix of a loop body function
*)

How does this command map to a WISL statment?

val stmt_kind_to_yojson : stmt_kind -> Yojson.Safe.t

val stmt_kind_of_yojson : Yojson.Safe.t -> stmt_kind Ppx_deriving_yojson_runtime.error_or

val pp_stmt_kind : Ppx_deriving_runtime.Format.formatter -> stmt_kind -> Ppx_deriving_runtime.unit

val show_stmt_kind : stmt_kind -> Ppx_deriving_runtime.string

type t = {

origin_loc : Gillian.Utils.Location.t option;
(*
Better not to know what this is for
*)

origin_id : int option;
(*
Origin Id, that should be abstracted away
*)

loop_info : string list;

stmt_kind : stmt_kind;

branch_kind : WUtils.WBranchCase.kind option;

nest_kind : nest_kind option;

}

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or

val _ : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or

val make : ?origin_loc:Gillian.Utils.Location.t -> ?origin_id:int -> ?loop_info:string list -> ?stmt_kind:stmt_kind -> ?branch_kind:WUtils.WBranchCase.kind -> ?nest_kind:nest_kind -> unit -> t

val make_multi : ?origin_loc:Gillian.Utils.Location.t -> ?origin_id:int -> ?loop_info:string list -> ?nest_kind:nest_kind -> ?is_return:bool -> unit -> t * t

val make_basic : ?origin_loc:Gillian.Utils.Location.t -> ?loop_info:string list -> unit -> t

val get_origin_loc : t -> Gillian.Utils.Location.t option

val get_loop_info : t -> string list

val set_loop_info : string list -> t -> t

val is_hidden : t -> bool